Moved views/settings to ThemeSettingsDialog
